feat: move all non-contract methods types into separate file
This will allow the JSON generator to only pull in the referenced types as needed, which will reduce the size of JSON files.
