+++ examples/counter/index.schema.json
@@ -1,19 +1,9 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
   "definitions": {
-    "AccountId": {
-      "maxLength": 64,
-      "minLength": 2,
-      "pattern": "^(([a-z\\d]+[-_])*[a-z\\d]+\\.)*([a-z\\d]+[-_])*[a-z\\d]+$",
-      "type": "string"
-    },
     "Balance": {
       "$ref": "#/definitions/U128",
       "description": "Balance is a type for storing amounts of tokens, specified in yoctoNEAR."
-    },
-    "Base64VecU8": {
-      "description": "base64 string.",
-      "type": "string"
     },
     "Decrement": {
       "additionalProperties": false,
@@ -50,14 +40,6 @@
     },
     "Decrement__Result": {
       "type": "null"
-    },
-    "Duration": {
-      "$ref": "#/definitions/u64",
-      "description": "Raw type for duration in nanoseconds"
-    },
-    "Gas": {
-      "$ref": "#/definitions/u64",
-      "description": "Represents the amount of NEAR tokens in \"gas units\" which are used to fund transactions."
     },
     "GetNum": {
       "additionalProperties": false,
@@ -113,10 +95,6 @@
     "Increment__Result": {
       "type": "null"
     },
-    "PublicKey": {
-      "description": "Public key in a binary format with base58 string serialization with human-readable curve. The key types currently supported are `secp256k1` and `ed25519`.\n\nEd25519 public keys accepted are 32 bytes and secp256k1 keys are the uncompressed 64 format.",
-      "type": "string"
-    },
     "Reset": {
       "additionalProperties": false,
       "contractMethod": "change",
@@ -153,14 +131,6 @@
     "Reset__Result": {
       "type": "null"
     },
-    "StorageUsage": {
-      "$ref": "#/definitions/u64",
-      "description": "StorageUsage is used to count the amount of storage used by a contract."
-    },
-    "Timestamp": {
-      "$ref": "#/definitions/u64",
-      "description": "Raw type for timestamp in nanoseconds"
-    },
     "U128": {
       "description": "String representation of a u128-bit integer",
       "pattern": "^[0-9]+$",
@@ -169,11 +139,6 @@
     "i8": {
       "maximum": 127,
       "minimum": -128,
-      "type": "integer"
-    },
-    "u64": {
-      "maximum": 18446744073709552000,
-      "minimum": 0,
       "type": "integer"
     }
   }
