+++ examples/counter/index.wit
@@ -23,17 +23,17 @@ get-num: function() -> s8
 ///  ```bash
 ///  near call counter.YOU.testnet increment --accountId donation.YOU.testnet
 ///  ```
-///  change
+///  @change
 increment: function()
 ///  Decrement (subtract from) the counter.
 /// 
 ///  ```bash
 ///  near call counter.YOU.testnet decrement --accountId donation.YOU.testnet
 ///  ```
-///  change
+///  @change
 decrement: function()
 ///  Reset to zero.
-///  change
+///  @change
 reset: function()// This is a generated file by witgen (https://github.com/bnjjj/witgen), please do not edit yourself, you can generate a new one thanks to cargo witgen generate command. (witme v0.2.0) 
 
 ///  StorageUsage is used to count the amount of storage used by a contract.
