+++ examples/rust-status-message-advanced/index.wit
@@ -13,7 +13,7 @@ record message {
 get-message: function() -> message
 
 ///  A change call to set the message
-///  change
+///  @change
 set-message: function(message: message)// This is a generated file by witgen (https://github.com/bnjjj/witgen), please do not edit yourself, you can generate a new one thanks to cargo witgen generate command. (witme v0.2.0) 
 
 ///  StorageUsage is used to count the amount of storage used by a contract.
